requirements.txt
# dependencies
mlflow
google-cloud-storage
pandas
scikit-learn
luigi
tensorflow